Isolate GameStateProvider in a separate file
