+++ src/SceneView.tsx
@@ -1,28 +1,25 @@
-import { observer } from 'mobx-react'
 import { useContext } from 'react'
-import { GameStateContext } from './App'
 import MomentView from './MomentView'
 import { Scene, SceneMomentType } from './scene-engine'
+import { GameStateContext } from './state'
 
 interface SceneViewProps {
   scene: Scene
+  sceneIndex: number
 }
 
 function SceneView(props: SceneViewProps) {
   const gameState = useContext(GameStateContext)
 
-  const { scene } = props
-  const moment = scene[gameState.sceneIndex]
+  const { scene, sceneIndex } = props
+  const moment = scene[sceneIndex]
 
   if (moment.momentType == SceneMomentType.JUMP) {
     gameState.jump(moment.destination)
+    return <></>
   } else {
-    return (
-      <div>
-        <MomentView moment={moment} />
-      </div>
-    )
+    return <MomentView moment={moment} />
   }
 }
 
-export default observer(SceneView)
+export default SceneView
